add line count to output

+++ spatial_experiment/learning_curve.py
@@ -4,6 +4,7 @@ import jiant.proj.main.export_model as export_model
 import jiant.proj.main.runscript as main_runscript
 import jiant.utils.python.io as py_io
 import jiant.utils.display as display
+import shutil
 import os
 
 jiant_run_config = configurator.SimpleAPIMultiTaskConfigurator(
@@ -25,6 +26,7 @@ export_model.export_model(
     hf_pretrained_model_name_or_path="bert-base-uncased",
     output_base_path="./spatial_experiment/models/bert",
 )
+
 
 def execute_training_loop():
     tokenize_and_cache.main(tokenize_and_cache.RunConfiguration(
@@ -49,6 +51,11 @@ def execute_training_loop():
     main_runscript.run_loop(run_args)
 
 
+train_path = 'spatial_experiment/tasks/data/spatial/train.jsonl'
+val_path = 'spatial_experiment/tasks/data/spatial/val.jsonl'
+test_path = 'spatial_experiment/tasks/data/spatial/test.jsonl'
+tr_vl_te = [train_path, val_path, test_path]
+
 d = '../spatial-commonsense/data/baseline'
 baseline_path = [os.path.join(d, o) for o in os.listdir(d) if os.path.isdir(os.path.join(d, o))]
 
@@ -58,22 +65,33 @@ for dir in baseline_path:
         f.write('name,acc,f1,recall,precision\n')
     print('wrote:', dir_name + '.csv')
     split_paths = [os.path.join(dir, o) for o in os.listdir(dir) if os.path.isdir(os.path.join(dir, o))]
+    for path in tr_vl_te:
+        if os.path.isfile(path):
+            os.remove(path)
+    shutil.copyfile(os.path.join(dir, 'train.jsonl'), train_path)
+    shutil.copyfile(os.path.join(dir, 'val.jsonl'), val_path)
+    shutil.copyfile(os.path.join(dir, 'test.jsonl'), test_path)
     execute_training_loop()
     with open('output.csv', 'r') as f:
         output = f.read()
-    max_test_path = os.path.join(dir_name, 'test.jsonl')
-    count = len(open(max_test_path).readlines())
+    max_test_path = os.path.join(dir, 'test.jsonl')
+    count = sum(1 for _ in open(max_test_path, encoding='utf8'))
     print(f'Number of rows in {max_test_path}: {count}')
     with open(dir_name + '.csv', 'a') as f:
         f.write(f'{str(count)},{output}')
     print(f'wrote: {str(count)},{output.strip()}')
-    os.remove('output.csv')
+    if os.path.isfile('output.csv'):
+        os.remove('output.csv')
     for split_dir in split_paths:
         split_dir_name = split_dir.split('/')[-1]
+        if os.path.isfile(train_path):
+            os.remove(train_path)
+        shutil.copyfile(os.path.join(split_dir, 'train.jsonl'), train_path)
         execute_training_loop()
         with open('output.csv', 'r') as f:
             output = f.read()
         with open(dir_name + '.csv', 'a') as f:
             f.write(f'{split_dir_name},{output}')
         print(f'wrote: {split_dir_name},{output.strip()}')
-        os.remove('output.csv')
+        if os.path.isfile('output.csv'):
+            os.remove('output.csv')
